+++ frontend/src/modals/EsiUpdate.tsx
@@ -1,17 +1,15 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { Button, Grid, IconButton, Link, TextField, Typography } from "@material-ui/core";
-import CloseIcon from '@material-ui/icons/Close';
+import { Button, Grid, Link, TextField, Typography } from "@material-ui/core";
 import axios from 'axios';
 import { GlobalDataContext } from '../GlobalDataContext';
-import { ResponseGatesUpdated, ResponseGates, ResponseMessage } from '../response';
+import { ResponseGatesUpdated, ResponseMessage } from '../response';
 import { dateFormat } from "../date";
 
-export default function UpdateGates() {
+export default function EsiUpdate() {
   const { t } = useTranslation();
   const globalData = useContext(GlobalDataContext);
   const [gatesUpdated, setGatesUpdated] = useState('');
-  const [gatesResult, setGatesResult] = useState<Array<string>|null>(null);
   const [searchTerm, setSearchTerm] = useState('»');
   const [searchResult, setSearchResult] = useState('');
   const [submitDisabled, setSubmitDisabled] = useState(false);
@@ -26,30 +24,9 @@ export default function UpdateGates() {
     }).catch(() => {});
   }, [globalData.domain, t]);
 
-  const fetchGates = (event: React.MouseEvent<HTMLButtonElement>) => {
-    const button = event.currentTarget;
-    button.disabled = true;
-    setGatesResult(null);
-    axios.get<ResponseGates>(`${globalData.domain}/api/gates/fetch`).then(response => {
-      let gates = [];
-      if (response.data.code) {
-        gates.push(t(`responseCode.${response.data.code}`));
-      } else {
-        for (let i = 0; i < response.data.ansiblexes.length; i++) {
-          gates.push(response.data.ansiblexes[i].name);
-        }
-      }
-      setGatesResult(gates);
-    }).catch(() => {
-      setGatesResult([t('app.error')]);
-    }).then(() => {
-      button.disabled = false;
-    });
-  };
-
   const search = () => {
     if (searchTerm.length === 0) {
-      setSearchResult(t('updateGates.input-required'));
+      setSearchResult(t('esiUpdate.input-required'));
       return;
     }
     setSearchResult('');
@@ -75,9 +52,8 @@ export default function UpdateGates() {
       <Grid container spacing={2}>
         <Grid item xs={12}>
           <Typography variant="body2">
-            <Trans i18nKey="updateGates.intro">
-              %
-              <Link target="_blank" rel="noopener noreferrer"
+            <Trans i18nKey="esiUpdate.intro">
+              %<Link target="_blank" rel="noopener noreferrer"
                 href="https://developers.eveonline.com/blog/article/the-esi-api-is-a-shared-resource-do-not-abuse-it"
               >%</Link>
               %<Link href="https://github.com/esi/esi-issues/issues/1185"
@@ -86,45 +62,23 @@ export default function UpdateGates() {
           </Typography>
           <br/>
           <Typography variant="body2">
-            {t('updateGates.update-hint')}<br/>
-            {t('updateGates.last-update')} {gatesUpdated}
+            {t('esiUpdate.update-hint')}<br/>
+            {t('esiUpdate.last-update')} {gatesUpdated}
           </Typography>
         </Grid>
         <Grid item xs={12}>
           <TextField
             required
             variant="filled"
-            label={t('updateGates.esi-search-label')}
-            helperText={t('updateGates.esi-search-help')}
+            label={t('esiUpdate.esi-search-label')}
+            helperText={t('esiUpdate.esi-search-help')}
             onChange={e => setSearchTerm(e.target.value)}
             style={{verticalAlign: 'baseline'}}
             defaultValue="»"
           />
           {' '}
-          <Button variant="contained"  disabled={submitDisabled}
-                  onClick={search}>{t('updateGates.submit')}</Button>
-
+          <Button variant="contained" disabled={submitDisabled} onClick={search}>{t('esiUpdate.submit')}</Button>
           <div>{searchResult}</div>
-        </Grid>
-        <Grid item xs={12}>
-          <br/>
-          <Button size={"small"} variant="contained" disableElevation
-                  onClick={fetchGates}>{t('updateGates.show-gates')}</Button>
-          {' '}
-          {gatesResult !== null &&
-          <span>
-            <IconButton size="small" onClick={() => setGatesResult(null)}><CloseIcon /></IconButton>
-            {' '}
-            {t('updateGates.num-gates', {number: gatesResult.length})}
-          </span>
-          }
-          {gatesResult !== null &&
-            <ul>
-              {gatesResult.map((value, index) => {
-                return <li key={index}>{value}<br/></li>
-              })}
-            </ul>
-          }
         </Grid>
       </Grid>
     </div>